preprocessing-image/requirements.txt
google-generativeai
pillow
pyodbc
matplotlib
python-magic
pandas
streamlit
pytesseract
opencv-python-headless
flask
request
python-dotenv
pandas
jsonify
uvicorn[standard]
textdistance
